- Move attachment functionality to Basecard
