+++ server/game/cards/02.4-TCT/PitTrap.js
@@ -8,7 +8,7 @@ class PitTrap extends DrawCard {
     }
 
     canPlayOn(card) {
-        return card.isAttacking() && super.canPlayOn();
+        return card.isAttacking() && super.canPlayOn(card);
     }
 }
